backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-cockroachdb==5.0
django-cors-headers==4.4.0
django-tidb==5.0.0
djangorestframework==3.15.2
mysqlclient==2.2.4
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
